+++ RELEASENOTES.md
@@ -49,6 +49,8 @@
 *   Cast extension
     *   Bump Cast SDK version to 21.2.0.
 *   IMA extension
+    *   Remove player listener of the `ImaServerSideAdInsertionMediaSource` on
+        the application thread to avoid threading issues.
     *   Bump IMA SDK version to 3.29.0.
 
 ### 1.0.0-beta03 (2022-11-22)